requirements.txt
streamlit
yfinance
pandas
matplotlib
pillow
sqlite3